remove proxy debugging code
Issue is the cli tool not setting a content-length
header. This isn't an issue when proxy-buffering is
enabled on the ingress since the entire body is buffered
and sent with a content-length header by nginx.
After disabling proxy-buffering, the proxied request
no longer has a content-length, so the warp::body
filters ignore the request.